refactor(v0): split provider tests for GH727
