+++ specs/GH727/tech.md
@@ -10,10 +10,10 @@ Link to `product.md`.
 
 ## Current Evidence
 
-At `origin/main@b1e5b85f`, 22 tracked Rust files remain over the U-16
-800-line ceiling. The current largest file is `src/core/observability/types.rs`
-at 856 lines. It is an observability type module where production definitions
-are below the ceiling and the oversized portion is the inline unit test suite.
+At `origin/main@a26d350c`, 21 tracked Rust files remain over the U-16
+800-line ceiling. The current largest file is `src/core/providers/v0/mod.rs`
+at 852 lines. It is a V0 provider module where production definitions are below
+the ceiling and the oversized portion is the inline unit test suite.
 
 ## Architecture Principles
 
@@ -40,59 +40,56 @@ are below the ceiling and the oversized portion is the inline unit test suite.
 | P4 | Utility modules | config helpers, net client utils, sync containers | focused utility tests plus concurrency/behavior checks where relevant |
 | P5 | Closure scan | all Rust files | full over-800 scan, final SpecRail update, final PR may close #727 |
 
-## Current Tranche: Observability Types Test Extraction
+## Current Tranche: V0 Provider Test Extraction
 
 ### Codebase Context
 
 | Area | Files | Current behavior | Why relevant |
 | --- | --- | --- | --- |
-| Type definitions | `src/core/observability/types.rs` | Defines metric values, observability log records, log levels, token/error details, alert state, trace spans, and span logs. | Production surface must keep the same names, fields, derives, serde behavior, conversion behavior, and observability re-export path. |
-| Inline tests | `src/core/observability/types.rs` | Contains metric value, log level, token usage, error details, alert, log record, trace span, span log, and workflow tests inline. | The test suite is what pushes the file over 800 lines. |
-| Extracted tests | `src/core/observability/types_tests.rs` | New path-backed test module loaded from `types.rs`. | Removes the oversized inline test block without changing runtime architecture. |
+| Provider definitions | `src/core/providers/v0/mod.rs` | Defines V0 config/model/provider, OpenAI-compatible parameter mapping, request/response transform, health check, cost calculation, and provider metadata. | Production surface must keep the same names, trait implementations, helper behavior, and module path. |
+| Inline tests | `src/core/providers/v0/mod.rs` | Contains config, ProviderConfig, model parsing, provider helper, metadata, cost, and parameter mapping tests inline. | The test suite is what pushes the file over 800 lines. |
+| Extracted tests | `src/core/providers/v0/tests.rs` | New path-backed test module loaded from `mod.rs`. | Removes the oversized inline test block without changing runtime architecture. |
 
 ### Design
 
-1. Keep `src/core/observability/types.rs` as the production observability type definition file.
-2. Preserve all public struct, enum, and type alias definitions in place, including fields,
-   derives, serde attributes, and `From` conversion implementations.
-3. Add only a `#[cfg(test)] #[path = "types_tests.rs"] mod tests;` declaration at the end
-   of `types.rs`.
-4. Move the original inline tests into `src/core/observability/types_tests.rs`.
-5. Keep the test module name as `core::observability::types::tests` so focused test filters and
-   historical paths continue to work.
-6. Do not edit observability logging, metrics, tracing, destinations, redaction, histogram, or module re-export files.
+1. Keep `src/core/providers/v0/mod.rs` as the production V0 provider module.
+2. Preserve all public struct, enum, helper function, and trait implementation definitions in place,
+   including fields, derives, defaults, request/response transforms, cost calculation, and error mapper selection.
+3. Add only a `#[cfg(test)] #[path = "tests.rs"] mod tests;` declaration at the end of `mod.rs`.
+4. Move the original inline tests into `src/core/providers/v0/tests.rs`.
+5. Keep the test module name as `core::providers::v0::tests` so focused test filters and historical paths continue to work.
+6. Do not edit `src/core/providers/v0/chat.rs` or any shared provider/runtime module.
 
 ## Product-to-Test Mapping
 
 | Product invariant | Implementation area | Verification |
 | --- | --- | --- |
-| P1 | `src/core/observability/types.rs` and `src/core/observability/mod.rs` | Public observability type exports stay on the original observability path. |
-| P2 | `src/core/observability/types.rs` | Root delegates tests with `#[path = "types_tests.rs"] mod tests;`. |
-| P3 | `src/core/observability/types_tests.rs` | Original inline observability type tests move without assertion changes. |
-| P4 | file size | `wc -l src/core/observability/types.rs src/core/observability/types_tests.rs` shows both files below 800. |
-| P5 | focused test suite | `cargo test core::observability::types --lib --all-features` runs the moved tests. |
-| P6 | queue count | tracked-file scan shows the remaining queue no longer includes `src/core/observability/types.rs`. |
+| P1 | `src/core/providers/v0/mod.rs` | Public V0 provider definitions stay in the original module path. |
+| P2 | `src/core/providers/v0/mod.rs` | Root delegates tests with `#[path = "tests.rs"] mod tests;`. |
+| P3 | `src/core/providers/v0/tests.rs` | Original inline V0 provider tests move without assertion changes. |
+| P4 | file size | `wc -l src/core/providers/v0/mod.rs src/core/providers/v0/tests.rs` shows both files below 800. |
+| P5 | focused test suite | `cargo test core::providers::v0 --lib --all-features` runs the moved tests. |
+| P6 | queue count | tracked-file scan shows the remaining queue no longer includes `src/core/providers/v0/mod.rs`. |
 
 ## Risks
 
-- Test extraction must not rename the module to a path that changes focused filters from
-  `core::observability::types::tests`.
-- The extracted test file relies on `use super::*` to keep access to private log-level conversion helpers and imported `HashMap`/`Utc` names used by the original inline tests.
-- Production observability type definitions should not be split in this tranche because they are already below 800 lines.
+- Test extraction must not rename the module to a path that changes focused filters from `core::providers::v0::tests`.
+- The extracted test file relies on `use super::*` to keep access to private helper methods like `get_endpoint` and `create_headers`.
+- Production V0 provider definitions should not be split in this tranche because they are already below 800 lines.
 
 ## Test Plan
 
 - [ ] `cargo fmt --all -- --check`
 - [ ] `git diff --check`
 - [ ] `python3 /Users/apple/Desktop/code/AI/tool/specrail/checks/check_workflow.py --repo /Users/apple/Desktop/code/AI/tool/specrail --spec-dir "$PWD/specs/GH727"`
-- [ ] `wc -l src/core/observability/types.rs src/core/observability/types_tests.rs`
-- [ ] `cargo test core::observability::types --lib --all-features`
+- [ ] `wc -l src/core/providers/v0/mod.rs src/core/providers/v0/tests.rs`
+- [ ] `cargo test core::providers::v0 --lib --all-features`
 - [ ] `cargo check --lib --all-features`
 - [ ] `cargo check --all-features --locked`
 - [ ] `cargo check`
 
 ## Rollback
 
-Move the observability type tests back into `src/core/observability/types.rs`
-and revert the `specs/GH727` edits. No schema, persistence, or runtime behavior
-changes are involved.
+Move the V0 provider tests back into `src/core/providers/v0/mod.rs` and revert
+the `specs/GH727` edits. No schema, persistence, or runtime behavior changes
+are involved.
